--- client/src/components/Admin.vue
@@ -1,63 +1,0 @@
-<template>
-  <div>
-    <vue-good-table
-      :columns="columns"
-      :rows="rows"/>
-  </div>
-</template>
-
-<script>
-export default {
-  name: 'my-component',
-  data(){
-    return {
-      columns: [
-        {
-          label: 'Name',
-          field: 'name',
-        },
-        {
-          label: 'Age',
-          field: 'age',
-          type: 'number',
-        },
-        {
-          label: 'Created On',
-          field: 'createdAt',
-          type: 'date',
-          dateInputFormat: 'yyyy-MM-dd',
-          dateOutputFormat: 'MMM Do yy',
-        },
-        {
-          label: 'Percent',
-          field: 'score',
-          type: 'percentage',
-        },
-      ],
-      rows: [
-        { id:1, name:"John", age: 20, createdAt: '2011-10-31',score: 0.03343 },
-        { id:2, name:"Jane", age: 24, createdAt: '2011-10-31', score: 0.03343 },
-        { id:3, name:"Susan", age: 16, createdAt: '2011-10-30', score: 0.03343 },
-        { id:4, name:"Chris", age: 55, createdAt: '2011-10-11', score: 0.03343 },
-        { id:5, name:"Dan", age: 40, createdAt: '2011-10-21', score: 0.03343 },
-        { id:6, name:"John", age: 20, createdAt: '2011-10-31', score: 0.03343 },
-      ],
-    };
-  },
-  created: function() {
-    this.getAllData()
-  },
-  methods: {
-    getAllData() {
-      this.axios.get("/students")
-      .then(response => {
-        if (response.status != 200) {
-          console.log("レスポンスエラー")
-        } else {
-          console.log(response.data)
-        }
-      })
-    }
-  },
-};
-</script>
