+++ client/src/views/Home.vue
@@ -1,18 +1,18 @@
 <template>
   <div class="home">
     <img alt="Vue logo" src="../assets/logo.png">
-    <Admin msg="Welcome to Your Vue.js App"/>
+    <bill-table msg="Welcome to Your Vue.js App"/>
   </div>
 </template>
 
 <script>
 // @ is an alias to /src
-import Admin from '@/components/Admin.vue'
+import BillTable from '@/components/BillTable.vue'
 
 export default {
   name: 'Home',
   components: {
-    Admin
+    BillTable
   }
 }
 </script>
